refactor(workspace): split attention provider
Moves generic workspace attention blockers out of WorkspaceProvider into a dedicated front/attention module so follow-up integration can consume blockers without provider coupling.
