+++ packages/workspace/src/front/attention/index.ts
@@ -1,0 +1,8 @@
+export {
+  WorkspaceAttentionProvider,
+  useWorkspaceAttention,
+} from "./WorkspaceAttentionProvider"
+export type {
+  WorkspaceAttentionBlocker,
+  WorkspaceAttentionContextValue,
+} from "./WorkspaceAttentionProvider"
